post compilation testing and minimal optimization
The start node is calculated so that the interpreter does not need to
step through all the subroutines.

The 'global' and 'local' commands are checked that they are allowed to
appear only before the first executable command of a subroutine. Also
the command 'local' should not appear out of a subroutine.

Commands 'use' and 'method' that declare global names to access Java
methods and classes are checked not to appear inside a subroutine. This
is because they are global and listing them inside a subroutine would
be misleading.