refactoring and comment polish
Now it is guaranteed that there is a next in the hook chain, and thus
there is no need to check it for the 3rd party hook developers. This is
because before all hook object a null hook object is registered.

The hook functions are now called not only when function call is
executed but also when an objects method is called.