command line
Old command line class remained in the source for reference purposes
and to allow it being executed with special command line. The new
command line class is CommandLineExtended. pom also references this new
class. The new command line class uses the native interface and allows
you to specify the extension classes that will be registered for the
interpreter to be executed.